fix: 李宏博　Task.25の品質向上
SQL文とMapper変数再構成

+++ trial_source/src/main/java/jp/co/sss/lms/mapper/TStudentAttendanceMapper.java
@@ -74,6 +74,6 @@ public interface TStudentAttendanceMapper {
 	 * @param trainingDate
 	 * @return 件数
 	 */
-	Integer notEnterCount(Integer lmsUserId,Integer deleteFlg,String trainingDate);
+	Integer notEnterCount(Integer lmsUserId,short deleteFlg,String trainingDate);
 
 }
